requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dm-haiku==0.0.12
dm-tree==0.1.8
dpath==2.1.6
dulwich==0.22.1
dvc==2.58.2
dvc-data==0.51.0
dvc-http==2.32.0
dvc-objects==0.25.0
dvc-render==0.7.0
dvc-studio-client==0.20.0
dvc-task==0.4.0
etils==1.8.0
executing==2.0.1
filelock==3.13.4
flatten-dict==0.4.2
flax==0.8.2
flufl.lock==8.1.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.0
grandalf==0.8
hydra-core==1.3.2
idna==3.7
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.23.0
iterative-telemetry==0.0.8
jax==0.4.26
jaxlib==0.4.26
jedi==0.19.1
jmp==0.0.4
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kombu==5.3.7
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
msgpack==1.0.8
multidict==6.0.5
nanotime==0.5.2
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.10
orjson==3.10.1
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pdoc3==0.10.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==3.11.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydot==2.0.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.2
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.13.0
scmrepo==1.4.1
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sqltrie==0.11.0
-e git+ssh://git@github.com/jiaxi98/neural-MFG.git@67031ff55cbe05f13a1066fd7e4f2907197274de#egg=src
stack-data==0.6.3
tabulate==0.9.0
tensorflow-datasets==4.9.4
tensorflow-metadata==1.15.0
tensorstore==0.1.57
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.18.1
